Invoke `cpu_loop()` to actually use multiple threads + Formatting tweaks
The function `cpu_loop()` was being invoked with only a single
thread. This commit changes the main file to be run with a single
argument representing the number of threads to be used. This is
used to invoke `cpu_loop()`, and is used to set the `num_threads`
clause of the OpenMP `#pragma omp parallel` inside the function.

Additionally, various parts of the logic in the file have been
tweaked to use named constants instead of the bare integers used
previously. The states of the MESI protocol, as well as the possible
instruction types, have been named using `enum`.

Apart from this, parts using dereferences with offsets (e.g. `*(x+i)`)
now use the corresponding syntactic sugar (e.g. `x[i]`) instead, for
easier reading.